// src/components/mixin/query-type.ts
enum QueryType {
  Point = 'Point',
  Circle = 'Circle',
  Rectangle = 'Rectangle',
  Polygon = 'Polygon',
  LineString = 'LineString',
  PickModel = 'PickModel',
  Cube = 'Cube',
  MultiPolygon = 'MultiPolygon',
  MyInputString = 'MyInputString',
  MyInputPolygon = 'MyInputPolygon',
  MyUpload = 'MyUpload',
  MyRegion = 'MyRegion',
}

export default QueryType